Split basically done
