+++ nexus/src/app/disk.rs
@@ -5,15 +5,17 @@
 use crate::authz;
 use crate::context::OpContext;
 use crate::db;
+use crate::db::lookup::LookupPath;
 use omicron_common::api::external::Error;
 use omicron_common::api::internal::nexus::DiskRuntimeState;
 use sled_agent_client::Client as SledAgentClient;
 use std::sync::Arc;
+use uuid::Uuid;
 
 impl super::Nexus {
     /// Modifies the runtime state of the Disk as requested.  This generally
     /// means attaching or detaching the disk.
-    pub (crate) async fn disk_set_runtime(
+    pub(crate) async fn disk_set_runtime(
         &self,
         opctx: &OpContext,
         authz_disk: &authz::Disk,
@@ -49,4 +51,60 @@ impl super::Nexus {
             .map(|_| ())
     }
 
+    pub async fn notify_disk_updated(
+        &self,
+        opctx: &OpContext,
+        id: Uuid,
+        new_state: &DiskRuntimeState,
+    ) -> Result<(), Error> {
+        let log = &self.log;
+        let (.., authz_disk) = LookupPath::new(&opctx, &self.db_datastore)
+            .disk_id(id)
+            .lookup_for(authz::Action::Modify)
+            .await?;
+
+        let result = self
+            .db_datastore
+            .disk_update_runtime(opctx, &authz_disk, &new_state.clone().into())
+            .await;
+
+        // TODO-cleanup commonize with notify_instance_updated()
+        match result {
+            Ok(true) => {
+                info!(log, "disk updated by sled agent";
+                    "disk_id" => %id,
+                    "new_state" => ?new_state);
+                Ok(())
+            }
+
+            Ok(false) => {
+                info!(log, "disk update from sled agent ignored (old)";
+                    "disk_id" => %id);
+                Ok(())
+            }
+
+            // If the disk doesn't exist, swallow the error -- there's
+            // nothing to do here.
+            // TODO-robustness This could only be possible if we've removed a
+            // disk from the datastore altogether.  When would we do that?
+            // We don't want to do it as soon as something's destroyed, I think,
+            // and in that case, we'd need some async task for cleaning these
+            // up.
+            Err(Error::ObjectNotFound { .. }) => {
+                warn!(log, "non-existent disk updated by sled agent";
+                    "instance_id" => %id,
+                    "new_state" => ?new_state);
+                Ok(())
+            }
+
+            // If the datastore is unavailable, propagate that to the caller.
+            Err(error) => {
+                warn!(log, "failed to update disk from sled agent";
+                    "disk_id" => %id,
+                    "new_state" => ?new_state,
+                    "error" => ?error);
+                Err(error)
+            }
+        }
+    }
 }
