+++ nexus/src/app/mod.rs
@@ -6,59 +6,22 @@
 
 use crate::authn;
 use crate::authz;
-use crate::authz::OrganizationRoles;
 use crate::config;
 use crate::context::OpContext;
 use crate::db;
-use crate::db::identity::Asset;
 use crate::db::lookup::LookupPath;
 use crate::db::model::Name;
-use crate::db::model::UpdateArtifactKind;
-use crate::external_api::shared;
-use crate::internal_api::params::OximeterInfo;
 use crate::populate::populate_start;
 use crate::populate::PopulateStatus;
 use crate::saga_interface::SagaContext;
 use crate::sagas;
 use anyhow::anyhow;
-use anyhow::Context;
-use futures::StreamExt;
-use hex;
-use omicron_common::api::external;
-use omicron_common::api::external::CreateResult;
 use omicron_common::api::external::DataPageParams;
-use omicron_common::api::external::DeleteResult;
 use omicron_common::api::external::Error;
-use omicron_common::api::external::ListResult;
 use omicron_common::api::external::ListResultVec;
 use omicron_common::api::external::LookupResult;
-use omicron_common::api::external::LookupType;
-use omicron_common::api::external::PaginationOrder;
-use omicron_common::api::external::ResourceType;
-use omicron_common::api::external::UpdateResult;
-use omicron_common::api::internal::nexus;
-use omicron_common::api::internal::nexus::DiskRuntimeState;
-use omicron_common::api::internal::nexus::UpdateArtifact;
-use omicron_common::backoff;
-use omicron_common::bail_unless;
-use oximeter_client::Client as OximeterClient;
-use oximeter_db::TimeseriesSchema;
-use oximeter_db::TimeseriesSchemaPaginationParams;
-use oximeter_producer::register;
-use rand::{rngs::StdRng, Rng, RngCore, SeedableRng};
-use ring::digest;
 use slog::Logger;
-use std::convert::{TryFrom, TryInto};
-use std::net::SocketAddr;
-use std::num::NonZeroU32;
-use std::path::Path;
 use std::sync::Arc;
-use std::time::Duration;
-use steno::SagaId;
-use steno::SagaResultOk;
-use steno::SagaTemplate;
-use steno::SagaType;
-use tokio::io::AsyncWriteExt;
 use uuid::Uuid;
 
 // The implementation of Nexus is large, and split into a number of submodules
@@ -69,26 +32,28 @@ mod image;
 mod instance;
 mod network_interface;
 mod organization;
+mod oximeter;
 mod project;
 mod rack;
-mod ssh_key;
+mod saga;
+mod session;
 mod silo;
 mod sled;
+mod ssh_key;
 // TODO: We should strongly consider moving all tests which depend on these
 // interfaces into unit tests, rather than integration tests.
 //
 // This would enable us to conditionally compile the test interfaces out of
 // the non-test builds, with #[cfg(test)]
 pub mod test_interfaces;
+mod update;
 mod vpc;
-mod vpc_subnet;
 mod vpc_router;
+mod vpc_subnet;
 mod zpool;
 
 // TODO: When referring to API types, we should try to include
 // the prefix unless it is unambiguous.
-
-pub static BASE_ARTIFACT_DIR: &str = "/var/tmp/oxide_artifacts";
 
 pub(crate) const MAX_DISKS_PER_INSTANCE: u32 = 8;
 
@@ -445,226 +410,8 @@ impl Nexus {
         }
     }
 
-    /// Insert a new record of an Oximeter collector server.
-    pub async fn upsert_oximeter_collector(
-        &self,
-        oximeter_info: &OximeterInfo,
-    ) -> Result<(), Error> {
-        // Insert the Oximeter instance into the DB. Note that this _updates_ the record,
-        // specifically, the time_modified, ip, and port columns, if the instance has already been
-        // registered.
-        let db_info = db::model::OximeterInfo::new(&oximeter_info);
-        self.db_datastore.oximeter_create(&db_info).await?;
-        info!(
-            self.log,
-            "registered new oximeter metric collection server";
-            "collector_id" => ?oximeter_info.collector_id,
-            "address" => oximeter_info.address,
-        );
-
-        // Regardless, notify the collector of any assigned metric producers. This should be empty
-        // if this Oximeter collector is registering for the first time, but may not be if the
-        // service is re-registering after failure.
-        let pagparams = DataPageParams {
-            marker: None,
-            direction: PaginationOrder::Ascending,
-            limit: std::num::NonZeroU32::new(100).unwrap(),
-        };
-        let producers = self
-            .db_datastore
-            .producers_list_by_oximeter_id(
-                oximeter_info.collector_id,
-                &pagparams,
-            )
-            .await?;
-        if !producers.is_empty() {
-            debug!(
-                self.log,
-                "registered oximeter collector that is already assigned producers, re-assigning them to the collector";
-                "n_producers" => producers.len(),
-                "collector_id" => ?oximeter_info.collector_id,
-            );
-            let client = self.build_oximeter_client(
-                &oximeter_info.collector_id,
-                oximeter_info.address,
-            );
-            for producer in producers.into_iter() {
-                let producer_info = oximeter_client::types::ProducerEndpoint {
-                    id: producer.id(),
-                    address: SocketAddr::new(
-                        producer.ip.ip(),
-                        producer.port.try_into().unwrap(),
-                    )
-                    .to_string(),
-                    base_route: producer.base_route,
-                    interval: oximeter_client::types::Duration::from(
-                        Duration::from_secs_f64(producer.interval),
-                    ),
-                };
-                client
-                    .producers_post(&producer_info)
-                    .await
-                    .map_err(Error::from)?;
-            }
-        }
-        Ok(())
-    }
-
-    /// Register as a metric producer with the oximeter metric collection server.
-    pub async fn register_as_producer(&self, address: SocketAddr) {
-        let producer_endpoint = nexus::ProducerEndpoint {
-            id: self.id,
-            address,
-            base_route: String::from("/metrics/collect"),
-            interval: Duration::from_secs(10),
-        };
-        let register = || async {
-            debug!(self.log, "registering nexus as metric producer");
-            register(address, &self.log, &producer_endpoint)
-                .await
-                .map_err(backoff::BackoffError::transient)
-        };
-        let log_registration_failure = |error, delay| {
-            warn!(
-                self.log,
-                "failed to register nexus as a metric producer, will retry in {:?}", delay;
-                "error_message" => ?error,
-            );
-        };
-        backoff::retry_notify(
-            backoff::internal_service_policy(),
-            register,
-            log_registration_failure,
-        ).await
-        .expect("expected an infinite retry loop registering nexus as a metric producer");
-    }
-
-    // Internal helper to build an Oximeter client from its ID and address (common data between
-    // model type and the API type).
-    fn build_oximeter_client(
-        &self,
-        id: &Uuid,
-        address: SocketAddr,
-    ) -> OximeterClient {
-        let client_log =
-            self.log.new(o!("oximeter-collector" => id.to_string()));
-        let client =
-            OximeterClient::new(&format!("http://{}", address), client_log);
-        info!(
-            self.log,
-            "registered oximeter collector client";
-            "id" => id.to_string(),
-        );
-        client
-    }
-
-    /// List all registered Oximeter collector instances.
-    pub async fn oximeter_list(
-        &self,
-        page_params: &DataPageParams<'_, Uuid>,
-    ) -> ListResultVec<db::model::OximeterInfo> {
-        self.db_datastore.oximeter_list(page_params).await
-    }
-
     pub fn datastore(&self) -> &Arc<db::DataStore> {
         &self.db_datastore
-    }
-
-    /// Given a saga template and parameters, create a new saga and execute it.
-    async fn execute_saga<P, S>(
-        self: &Arc<Self>,
-        saga_template: Arc<SagaTemplate<S>>,
-        template_name: &str,
-        saga_params: Arc<P>,
-    ) -> Result<SagaResultOk, Error>
-    where
-        S: SagaType<
-            ExecContextType = Arc<SagaContext>,
-            SagaParamsType = Arc<P>,
-        >,
-        // TODO-cleanup The bound `P: Serialize` should not be necessary because
-        // SagaParamsType must already impl Serialize.
-        P: serde::Serialize,
-    {
-        let saga_id = SagaId(Uuid::new_v4());
-        let saga_logger =
-            self.log.new(o!("template_name" => template_name.to_owned()));
-        let saga_context = Arc::new(Arc::new(SagaContext::new(
-            Arc::clone(self),
-            saga_logger,
-            Arc::clone(&self.authz),
-        )));
-        let future = self
-            .sec_client
-            .saga_create(
-                saga_id,
-                saga_context,
-                saga_template,
-                template_name.to_owned(),
-                saga_params,
-            )
-            .await
-            .context("creating saga")
-            .map_err(|error| {
-                // TODO-error This could be a service unavailable error,
-                // depending on the failure mode.  We need more information from
-                // Steno.
-                Error::internal_error(&format!("{:#}", error))
-            })?;
-
-        self.sec_client
-            .saga_start(saga_id)
-            .await
-            .context("starting saga")
-            .map_err(|error| Error::internal_error(&format!("{:#}", error)))?;
-
-        let result = future.await;
-        result.kind.map_err(|saga_error| {
-            saga_error.error_source.convert::<Error>().unwrap_or_else(|e| {
-                // TODO-error more context would be useful
-                Error::InternalError { internal_message: e.to_string() }
-            })
-        })
-    }
-
-    // Sagas
-
-    pub async fn sagas_list(
-        &self,
-        opctx: &OpContext,
-        pagparams: &DataPageParams<'_, Uuid>,
-    ) -> ListResult<external::Saga> {
-        // The endpoint we're serving only supports `ScanById`, which only
-        // supports an ascending scan.
-        bail_unless!(
-            pagparams.direction == dropshot::PaginationOrder::Ascending
-        );
-        opctx.authorize(authz::Action::Read, &authz::FLEET).await?;
-        let marker = pagparams.marker.map(|s| SagaId::from(*s));
-        let saga_list = self
-            .sec_client
-            .saga_list(marker, pagparams.limit)
-            .await
-            .into_iter()
-            .map(external::Saga::from)
-            .map(Ok);
-        Ok(futures::stream::iter(saga_list).boxed())
-    }
-
-    pub async fn saga_get(
-        &self,
-        opctx: &OpContext,
-        id: Uuid,
-    ) -> LookupResult<external::Saga> {
-        opctx.authorize(authz::Action::Read, &authz::FLEET).await?;
-        self.sec_client
-            .saga_get(steno::SagaId::from(id))
-            .await
-            .map(external::Saga::from)
-            .map(Ok)
-            .map_err(|_: ()| {
-                Error::not_found_by_id(ResourceType::SagaDbg, &id)
-            })?
     }
 
     // Built-in users
@@ -710,593 +457,6 @@ impl Nexus {
             .await?;
         Ok(db_role_builtin)
     }
-
-    // Internal control plane interfaces.
-
-    /// Invoked by a sled agent to publish an updated runtime state for an
-    /// Instance.
-    pub async fn notify_instance_updated(
-        &self,
-        id: &Uuid,
-        new_runtime_state: &nexus::InstanceRuntimeState,
-    ) -> Result<(), Error> {
-        let log = &self.log;
-
-        let result = self
-            .db_datastore
-            .instance_update_runtime(id, &(new_runtime_state.clone().into()))
-            .await;
-
-        match result {
-            Ok(true) => {
-                info!(log, "instance updated by sled agent";
-                    "instance_id" => %id,
-                    "propolis_id" => %new_runtime_state.propolis_uuid,
-                    "new_state" => %new_runtime_state.run_state);
-                Ok(())
-            }
-
-            Ok(false) => {
-                info!(log, "instance update from sled agent ignored (old)";
-                    "instance_id" => %id,
-                    "propolis_id" => %new_runtime_state.propolis_uuid,
-                    "requested_state" => %new_runtime_state.run_state);
-                Ok(())
-            }
-
-            // If the instance doesn't exist, swallow the error -- there's
-            // nothing to do here.
-            // TODO-robustness This could only be possible if we've removed an
-            // Instance from the datastore altogether.  When would we do that?
-            // We don't want to do it as soon as something's destroyed, I think,
-            // and in that case, we'd need some async task for cleaning these
-            // up.
-            Err(Error::ObjectNotFound { .. }) => {
-                warn!(log, "non-existent instance updated by sled agent";
-                    "instance_id" => %id,
-                    "new_state" => %new_runtime_state.run_state);
-                Ok(())
-            }
-
-            // If the datastore is unavailable, propagate that to the caller.
-            // TODO-robustness Really this should be any _transient_ error.  How
-            // can we distinguish?  Maybe datastore should emit something
-            // different from Error with an Into<Error>.
-            Err(error) => {
-                warn!(log, "failed to update instance from sled agent";
-                    "instance_id" => %id,
-                    "new_state" => %new_runtime_state.run_state,
-                    "error" => ?error);
-                Err(error)
-            }
-        }
-    }
-
-    pub async fn notify_disk_updated(
-        &self,
-        opctx: &OpContext,
-        id: Uuid,
-        new_state: &DiskRuntimeState,
-    ) -> Result<(), Error> {
-        let log = &self.log;
-        let (.., authz_disk) = LookupPath::new(&opctx, &self.db_datastore)
-            .disk_id(id)
-            .lookup_for(authz::Action::Modify)
-            .await?;
-
-        let result = self
-            .db_datastore
-            .disk_update_runtime(opctx, &authz_disk, &new_state.clone().into())
-            .await;
-
-        // TODO-cleanup commonize with notify_instance_updated()
-        match result {
-            Ok(true) => {
-                info!(log, "disk updated by sled agent";
-                    "disk_id" => %id,
-                    "new_state" => ?new_state);
-                Ok(())
-            }
-
-            Ok(false) => {
-                info!(log, "disk update from sled agent ignored (old)";
-                    "disk_id" => %id);
-                Ok(())
-            }
-
-            // If the disk doesn't exist, swallow the error -- there's
-            // nothing to do here.
-            // TODO-robustness This could only be possible if we've removed a
-            // disk from the datastore altogether.  When would we do that?
-            // We don't want to do it as soon as something's destroyed, I think,
-            // and in that case, we'd need some async task for cleaning these
-            // up.
-            Err(Error::ObjectNotFound { .. }) => {
-                warn!(log, "non-existent disk updated by sled agent";
-                    "instance_id" => %id,
-                    "new_state" => ?new_state);
-                Ok(())
-            }
-
-            // If the datastore is unavailable, propagate that to the caller.
-            Err(error) => {
-                warn!(log, "failed to update disk from sled agent";
-                    "disk_id" => %id,
-                    "new_state" => ?new_state,
-                    "error" => ?error);
-                Err(error)
-            }
-        }
-    }
-
-    // Timeseries
-
-    /// List existing timeseries schema.
-    pub async fn timeseries_schema_list(
-        &self,
-        opctx: &OpContext,
-        pag_params: &TimeseriesSchemaPaginationParams,
-        limit: NonZeroU32,
-    ) -> Result<dropshot::ResultsPage<TimeseriesSchema>, Error> {
-        opctx.authorize(authz::Action::Read, &authz::FLEET).await?;
-        self.timeseries_client
-            .timeseries_schema_list(&pag_params.page, limit)
-            .await
-            .map_err(|e| match e {
-                oximeter_db::Error::DatabaseUnavailable(_) => {
-                    Error::ServiceUnavailable {
-                        internal_message: e.to_string(),
-                    }
-                }
-                _ => Error::InternalError { internal_message: e.to_string() },
-            })
-    }
-
-    /// Assign a newly-registered metric producer to an oximeter collector server.
-    pub async fn assign_producer(
-        &self,
-        producer_info: nexus::ProducerEndpoint,
-    ) -> Result<(), Error> {
-        let (collector, id) = self.next_collector().await?;
-        let db_info = db::model::ProducerEndpoint::new(&producer_info, id);
-        self.db_datastore.producer_endpoint_create(&db_info).await?;
-        collector
-            .producers_post(&oximeter_client::types::ProducerEndpoint::from(
-                &producer_info,
-            ))
-            .await
-            .map_err(Error::from)?;
-        info!(
-            self.log,
-            "assigned collector to new producer";
-            "producer_id" => ?producer_info.id,
-            "collector_id" => ?id,
-        );
-        Ok(())
-    }
-
-    /// Return an oximeter collector to assign a newly-registered producer
-    async fn next_collector(&self) -> Result<(OximeterClient, Uuid), Error> {
-        // TODO-robustness Replace with a real load-balancing strategy.
-        let page_params = DataPageParams {
-            marker: None,
-            direction: dropshot::PaginationOrder::Ascending,
-            limit: std::num::NonZeroU32::new(1).unwrap(),
-        };
-        let oxs = self.db_datastore.oximeter_list(&page_params).await?;
-        let info = oxs.first().ok_or_else(|| Error::ServiceUnavailable {
-            internal_message: String::from("no oximeter collectors available"),
-        })?;
-        let address =
-            SocketAddr::from((info.ip.ip(), info.port.try_into().unwrap()));
-        let id = info.id;
-        Ok((self.build_oximeter_client(&id, address), id))
-    }
-
-    pub async fn session_fetch(
-        &self,
-        opctx: &OpContext,
-        token: String,
-    ) -> LookupResult<authn::ConsoleSessionWithSiloId> {
-        let (.., db_console_session) =
-            LookupPath::new(opctx, &self.db_datastore)
-                .console_session_token(&token)
-                .fetch()
-                .await?;
-
-        let (.., db_silo_user) = LookupPath::new(opctx, &self.db_datastore)
-            .silo_user_id(db_console_session.silo_user_id)
-            .fetch()
-            .await?;
-
-        Ok(authn::ConsoleSessionWithSiloId {
-            console_session: db_console_session,
-            silo_id: db_silo_user.silo_id,
-        })
-    }
-
-    pub async fn session_create(
-        &self,
-        opctx: &OpContext,
-        user_id: Uuid,
-    ) -> CreateResult<db::model::ConsoleSession> {
-        if !self.login_allowed(opctx, user_id).await? {
-            return Err(Error::Unauthenticated {
-                internal_message: "User not allowed to login".to_string(),
-            });
-        }
-
-        let session =
-            db::model::ConsoleSession::new(generate_session_token(), user_id);
-
-        self.db_datastore.session_create(opctx, session).await
-    }
-
-    // update last_used to now
-    pub async fn session_update_last_used(
-        &self,
-        opctx: &OpContext,
-        token: &str,
-    ) -> UpdateResult<authn::ConsoleSessionWithSiloId> {
-        let authz_session = authz::ConsoleSession::new(
-            authz::FLEET,
-            token.to_string(),
-            LookupType::ByCompositeId(token.to_string()),
-        );
-        self.db_datastore.session_update_last_used(opctx, &authz_session).await
-    }
-
-    pub async fn session_hard_delete(
-        &self,
-        opctx: &OpContext,
-        token: &str,
-    ) -> DeleteResult {
-        let authz_session = authz::ConsoleSession::new(
-            authz::FLEET,
-            token.to_string(),
-            LookupType::ByCompositeId(token.to_string()),
-        );
-        self.db_datastore.session_hard_delete(opctx, &authz_session).await
-    }
-
-    fn tuf_base_url(&self) -> Option<String> {
-        self.updates_config.as_ref().map(|c| {
-            let rack = self.as_rack();
-            rack.tuf_base_url.unwrap_or_else(|| c.default_base_url.clone())
-        })
-    }
-
-    pub async fn updates_refresh_metadata(
-        &self,
-        opctx: &OpContext,
-    ) -> Result<(), Error> {
-        opctx.authorize(authz::Action::Modify, &authz::FLEET).await?;
-
-        let updates_config = self.updates_config.as_ref().ok_or_else(|| {
-            Error::InvalidRequest {
-                message: "updates system not configured".into(),
-            }
-        })?;
-        let base_url =
-            self.tuf_base_url().ok_or_else(|| Error::InvalidRequest {
-                message: "updates system not configured".into(),
-            })?;
-        let trusted_root = tokio::fs::read(&updates_config.trusted_root)
-            .await
-            .map_err(|e| Error::InternalError {
-                internal_message: format!(
-                    "error trying to read trusted root: {}",
-                    e
-                ),
-            })?;
-
-        let artifacts = tokio::task::spawn_blocking(move || {
-            crate::updates::read_artifacts(&trusted_root, base_url)
-        })
-        .await
-        .unwrap()
-        .map_err(|e| Error::InternalError {
-            internal_message: format!("error trying to refresh updates: {}", e),
-        })?;
-
-        // FIXME: if we hit an error in any of these database calls, the
-        // available artifact table will be out of sync with the current
-        // artifacts.json. can we do a transaction or something?
-
-        let mut current_version = None;
-        for artifact in &artifacts {
-            current_version = Some(artifact.targets_role_version);
-            self.db_datastore
-                .update_available_artifact_upsert(&opctx, artifact.clone())
-                .await?;
-        }
-
-        // ensure table is in sync with current copy of artifacts.json
-        if let Some(current_version) = current_version {
-            self.db_datastore
-                .update_available_artifact_hard_delete_outdated(
-                    &opctx,
-                    current_version,
-                )
-                .await?;
-        }
-
-        // demo-grade update logic: tell all sleds to apply all artifacts
-        for sled in self
-            .db_datastore
-            .sled_list(
-                &opctx,
-                &DataPageParams {
-                    marker: None,
-                    direction: PaginationOrder::Ascending,
-                    limit: NonZeroU32::new(100).unwrap(),
-                },
-            )
-            .await?
-        {
-            let client = self.sled_client(&sled.id()).await?;
-            for artifact in &artifacts {
-                info!(
-                    self.log,
-                    "telling sled {} to apply {}",
-                    sled.id(),
-                    artifact.target_name
-                );
-                client
-                    .update_artifact(
-                        &sled_agent_client::types::UpdateArtifact {
-                            name: artifact.name.clone(),
-                            version: artifact.version,
-                            kind: artifact.kind.0.into(),
-                        },
-                    )
-                    .await?;
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Downloads a file from within [`BASE_ARTIFACT_DIR`].
-    pub async fn download_artifact(
-        &self,
-        opctx: &OpContext,
-        artifact: UpdateArtifact,
-    ) -> Result<Vec<u8>, Error> {
-        let mut base_url =
-            self.tuf_base_url().ok_or_else(|| Error::InvalidRequest {
-                message: "updates system not configured".into(),
-            })?;
-        if !base_url.ends_with('/') {
-            base_url.push('/');
-        }
-
-        // We cache the artifact based on its checksum, so fetch that from the
-        // database.
-        let (.., artifact_entry) = LookupPath::new(opctx, &self.db_datastore)
-            .update_available_artifact_tuple(
-                &artifact.name,
-                artifact.version,
-                UpdateArtifactKind(artifact.kind),
-            )
-            .fetch()
-            .await?;
-        let filename = format!(
-            "{}.{}.{}-{}",
-            artifact_entry.target_sha256,
-            artifact.kind,
-            artifact.name,
-            artifact.version
-        );
-        let path = Path::new(BASE_ARTIFACT_DIR).join(&filename);
-
-        if !path.exists() {
-            // If the artifact doesn't exist, we should download it.
-            //
-            // TODO: There also exists the question of "when should we *remove*
-            // things from BASE_ARTIFACT_DIR", which we should also resolve.
-            // Demo-quality solution could be "destroy it on boot" or something?
-            // (we aren't doing that yet).
-            info!(self.log, "Accessing {} - needs to be downloaded", filename);
-            tokio::fs::create_dir_all(BASE_ARTIFACT_DIR).await.map_err(
-                |e| {
-                    Error::internal_error(&format!(
-                        "Failed to create artifacts directory: {}",
-                        e
-                    ))
-                },
-            )?;
-
-            let mut response = reqwest::get(format!(
-                "{}targets/{}.{}",
-                base_url,
-                artifact_entry.target_sha256,
-                artifact_entry.target_name
-            ))
-            .await
-            .map_err(|e| {
-                Error::internal_error(&format!(
-                    "Failed to fetch artifact: {}",
-                    e
-                ))
-            })?;
-
-            // To ensure another request isn't trying to use this target while we're downloading it
-            // or before we've verified it, write to a random path in the same directory, then move
-            // it to the correct path after verification.
-            let temp_path = path.with_file_name(format!(
-                ".{}.{:x}",
-                filename,
-                rand::thread_rng().gen::<u64>()
-            ));
-            let mut file =
-                tokio::fs::File::create(&temp_path).await.map_err(|e| {
-                    Error::internal_error(&format!(
-                        "Failed to create file: {}",
-                        e
-                    ))
-                })?;
-
-            let mut context = digest::Context::new(&digest::SHA256);
-            let mut length: i64 = 0;
-            while let Some(chunk) = response.chunk().await.map_err(|e| {
-                Error::internal_error(&format!(
-                    "Failed to read HTTP body: {}",
-                    e
-                ))
-            })? {
-                file.write_all(&chunk).await.map_err(|e| {
-                    Error::internal_error(&format!(
-                        "Failed to write to file: {}",
-                        e
-                    ))
-                })?;
-                context.update(&chunk);
-                length += i64::try_from(chunk.len()).unwrap();
-
-                if length > artifact_entry.target_length {
-                    return Err(Error::internal_error(&format!(
-                        "target {} is larger than expected",
-                        artifact_entry.target_name
-                    )));
-                }
-            }
-            drop(file);
-
-            if hex::encode(context.finish()) == artifact_entry.target_sha256
-                && length == artifact_entry.target_length
-            {
-                tokio::fs::rename(temp_path, &path).await.map_err(|e| {
-                    Error::internal_error(&format!(
-                        "Failed to rename file after verification: {}",
-                        e
-                    ))
-                })?
-            } else {
-                return Err(Error::internal_error(&format!(
-                    "failed to verify target {}",
-                    artifact_entry.target_name
-                )));
-            }
-
-            info!(
-                self.log,
-                "wrote {} to artifact dir", artifact_entry.target_name
-            );
-        } else {
-            info!(self.log, "Accessing {} - already exists", path.display());
-        }
-
-        // TODO: These artifacts could be quite large - we should figure out how to
-        // stream this file back instead of holding it entirely in-memory in a
-        // Vec<u8>.
-        //
-        // Options:
-        // - RFC 7233 - "Range Requests" (is this HTTP/1.1 only?)
-        // https://developer.mozilla.org/en-US/docs/Web/HTTP/Range_requests
-        // - "Roll our own". See:
-        // https://stackoverflow.com/questions/20969331/standard-method-for-http-partial-upload-resume-upload
-        let body = tokio::fs::read(&path).await.map_err(|e| {
-            Error::internal_error(&format!(
-                "Cannot read artifact from filesystem: {}",
-                e
-            ))
-        })?;
-        Ok(body)
-    }
-
-    async fn login_allowed(
-        &self,
-        opctx: &OpContext,
-        silo_user_id: Uuid,
-    ) -> Result<bool, Error> {
-        // Was this silo user deleted?
-        let fetch_result = LookupPath::new(opctx, &self.db_datastore)
-            .silo_user_id(silo_user_id)
-            .fetch()
-            .await;
-
-        match fetch_result {
-            Err(e) => {
-                match e {
-                    Error::ObjectNotFound { type_name: _, lookup_type: _ } => {
-                        // if the silo user was deleted, they're not allowed to
-                        // log in :)
-                        return Ok(false);
-                    }
-
-                    _ => {
-                        return Err(e);
-                    }
-                }
-            }
-
-            Ok(_) => {
-                // they're allowed
-            }
-        }
-
-        Ok(true)
-    }
-
-    pub async fn silo_user_fetch(
-        &self,
-        opctx: &OpContext,
-        silo_user_id: Uuid,
-    ) -> LookupResult<db::model::SiloUser> {
-        let (.., db_silo_user) = LookupPath::new(opctx, &self.datastore())
-            .silo_user_id(silo_user_id)
-            .fetch()
-            .await?;
-        Ok(db_silo_user)
-    }
-
-    // Role assignments
-
-    pub async fn organization_fetch_policy(
-        &self,
-        opctx: &OpContext,
-        organization_name: &Name,
-    ) -> LookupResult<shared::Policy<OrganizationRoles>> {
-        let (.., authz_org) = LookupPath::new(opctx, &self.db_datastore)
-            .organization_name(organization_name)
-            .lookup_for(authz::Action::ReadPolicy)
-            .await?;
-        let role_assignments = self
-            .db_datastore
-            .role_assignment_fetch_all(opctx, &authz_org)
-            .await?
-            .into_iter()
-            .map(|r| r.try_into().context("parsing database role assignment"))
-            .collect::<Result<Vec<_>, _>>()
-            .map_err(|error| Error::internal_error(&format!("{:#}", error)))?;
-        Ok(shared::Policy { role_assignments })
-    }
-
-    pub async fn organization_update_policy(
-        &self,
-        opctx: &OpContext,
-        organization_name: &Name,
-        policy: &shared::Policy<OrganizationRoles>,
-    ) -> UpdateResult<shared::Policy<OrganizationRoles>> {
-        let (.., authz_org) = LookupPath::new(opctx, &self.db_datastore)
-            .organization_name(organization_name)
-            .lookup_for(authz::Action::ModifyPolicy)
-            .await?;
-
-        let role_assignments = self
-            .db_datastore
-            .role_assignment_replace_all(
-                opctx,
-                &authz_org,
-                &policy.role_assignments,
-            )
-            .await?
-            .into_iter()
-            .map(|r| r.try_into())
-            .collect::<Result<Vec<_>, _>>()?;
-        Ok(shared::Policy { role_assignments })
-    }
 }
 
 /// For unimplemented endpoints, indicates whether the resource identified
@@ -1312,17 +472,4 @@ impl Nexus {
 pub enum Unimpl {
     Public,
     ProtectedLookup(Error),
-}
-
-fn generate_session_token() -> String {
-    // TODO: "If getrandom is unable to provide secure entropy this method will panic."
-    // Should we explicitly handle that?
-    // TODO: store generator somewhere so we don't reseed every time
-    let mut rng = StdRng::from_entropy();
-    // OWASP recommends at least 64 bits of entropy, OAuth 2 spec 128 minimum, 160 recommended
-    // 20 bytes = 160 bits of entropy
-    // TODO: the size should be a constant somewhere, maybe even in config?
-    let mut random_bytes: [u8; 20] = [0; 20];
-    rng.fill_bytes(&mut random_bytes);
-    hex::encode(random_bytes)
 }
