+++ nexus/src/app/organization.rs
@@ -3,14 +3,18 @@
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
 use crate::authz;
+use crate::authz::OrganizationRoles;
 use crate::context::OpContext;
 use crate::db;
 use crate::db::lookup::LookupPath;
 use crate::db::model::Name;
 use crate::external_api::params;
+use crate::external_api::shared;
+use anyhow::Context;
 use omicron_common::api::external::CreateResult;
 use omicron_common::api::external::DataPageParams;
 use omicron_common::api::external::DeleteResult;
+use omicron_common::api::external::Error;
 use omicron_common::api::external::ListResultVec;
 use omicron_common::api::external::LookupResult;
 use omicron_common::api::external::UpdateResult;
@@ -84,5 +88,52 @@ impl super::Nexus {
                 new_params.clone().into(),
             )
             .await
+    }
+
+    // Role assignments
+
+    pub async fn organization_fetch_policy(
+        &self,
+        opctx: &OpContext,
+        organization_name: &Name,
+    ) -> LookupResult<shared::Policy<OrganizationRoles>> {
+        let (.., authz_org) = LookupPath::new(opctx, &self.db_datastore)
+            .organization_name(organization_name)
+            .lookup_for(authz::Action::ReadPolicy)
+            .await?;
+        let role_assignments = self
+            .db_datastore
+            .role_assignment_fetch_all(opctx, &authz_org)
+            .await?
+            .into_iter()
+            .map(|r| r.try_into().context("parsing database role assignment"))
+            .collect::<Result<Vec<_>, _>>()
+            .map_err(|error| Error::internal_error(&format!("{:#}", error)))?;
+        Ok(shared::Policy { role_assignments })
+    }
+
+    pub async fn organization_update_policy(
+        &self,
+        opctx: &OpContext,
+        organization_name: &Name,
+        policy: &shared::Policy<OrganizationRoles>,
+    ) -> UpdateResult<shared::Policy<OrganizationRoles>> {
+        let (.., authz_org) = LookupPath::new(opctx, &self.db_datastore)
+            .organization_name(organization_name)
+            .lookup_for(authz::Action::ModifyPolicy)
+            .await?;
+
+        let role_assignments = self
+            .db_datastore
+            .role_assignment_replace_all(
+                opctx,
+                &authz_org,
+                &policy.role_assignments,
+            )
+            .await?
+            .into_iter()
+            .map(|r| r.try_into())
+            .collect::<Result<Vec<_>, _>>()?;
+        Ok(shared::Policy { role_assignments })
     }
 }
