+++ nexus/src/app/sled.rs
@@ -6,8 +6,8 @@ use crate::context::OpContext;
 use crate::db;
 use crate::db::identity::Asset;
 use crate::db::lookup::LookupPath;
-use omicron_common::api::external::Error;
 use omicron_common::api::external::DataPageParams;
+use omicron_common::api::external::Error;
 use omicron_common::api::external::ListResultVec;
 use omicron_common::api::external::LookupResult;
 use sled_agent_client::Client as SledAgentClient;
