+++ nexus/src/lib.rs
@@ -28,13 +28,13 @@ mod saga_interface;
 mod sagas;
 pub mod updates; // public for testing
 
+pub use app::test_interfaces::TestInterfaces;
+pub use app::Nexus;
 pub use config::Config;
 pub use context::ServerContext;
 pub use crucible_agent_client;
 use external_api::http_entrypoints::external_api;
 use internal_api::http_entrypoints::internal_api;
-pub use app::Nexus;
-pub use app::test_interfaces::TestInterfaces;
 use slog::Logger;
 use std::sync::Arc;
 use uuid::Uuid;
